requirements.txt
langchain
langchain-core
langchain-groq
langchain-google-genai
grandalf
python-dotenv
langgraph
pydantic
langchain-community
tavily-python
langgraph-checkpoint-sqlite
faiss-cpu
sentence-transformers